+++ test/example.js
@@ -15,9 +15,22 @@ describe('example', function () {
     describe('get example', function () {
 
         it('check types list', function (done) {
-            this.vm.getAllTypes().should.be.instanceof(Array);;
+            this.vm.getAllTypes().should.be.instanceof(Array);
             this.vm.getAllTypes().should.contain('string');
             this.vm.getAllTypes().should.contain('integer');
+            done();
+          });
+
+        it('check getExamples', function (done) {
+            var types = this.vm.getAllTypes();
+            for (var i = 0; i < types.length; i++) {
+              var example = this.vm.getExamples(types[i]);
+              example.should.be.instanceof(Array);
+              this.vm.validate(types[i], example[0], function (err) {
+                (err === null).should.be.true;
+              });
+            }
+
             done();
           });
 
@@ -33,8 +46,25 @@ describe('example', function () {
             done();
           });
 
+        it('check all', function (done) {
+            var types = this.vm.getAllTypes();
+            for (var i = 0; i < types.length; i++) {
+              var example = this.vm.getRandomExample(types[i]);
+              this.vm.validate(types[i], example, function (err) {
+                (err === null).should.be.true;
+              });
+            }
+
+            done();
+          });
+
         it('check notype', function (done) {
             (this.vm.getExample('notype') === undefined).should.be.true;
+            done();
+          });
+
+        it('check notype', function (done) {
+            (this.vm.getRandomExample('notype') === undefined).should.be.true;
             done();
           });